+++ contenido/includes/include.style_history.php
@@ -19,6 +19,11 @@ cInclude('includes', 'functions.lay.php');
 cInclude('includes', 'functions.file.php');
 cInclude('external', 'codemirror/class.codemirror.php');
 
+$readOnly = (getEffectiveSetting("client", "readonly", "false") == "true");
+if($readOnly) {
+    cRegistry::addWarningMessage(i18n("The administrator disbaled editing these files!"));
+}
+
 $sFileName = '';
 $sFileName = $_REQUEST['file'];
 
@@ -49,13 +54,13 @@ if (!$perm->have_perm_area_action($area, 'style_history_manage')) {
     $aFileInfo = $fileInfoCollection->getFileInformation($sFileName, $sTypeContent);
 
     // [action] => history_truncate delete all current history
-    if ($_POST['action'] == 'history_truncate') {
+    if ((!$readOnly) && $_POST['action'] == 'history_truncate') {
         $oVersionStyle = new cVersionFile($aFileInfo['idsfi'], $aFileInfo, $sFileName, $sTypeContent, $cfg, $cfgClient, $db, $client, $area, $frame);
         $bDeleteFile = $oVersionStyle->deleteFile();
         unset($oVersionStyle);
     }
 
-    if ($_POST['style_send'] == true && $_POST['stylecode'] != '' && $sFileName != '' && $aFileInfo['idsfi'] != '') { // save
+    if ((!$readOnly) && $_POST['style_send'] == true && $_POST['stylecode'] != '' && $sFileName != '' && $aFileInfo['idsfi'] != '') { // save
                                                                                                                       // button
                                                                                                                       // Get
                                                                                                                       // Post
@@ -98,7 +103,7 @@ if (!$perm->have_perm_area_action($area, 'style_history_manage')) {
         unset($oVersionStyle);
     }
 
-    if ($sFileName != '' && $aFileInfo['idsfi'] != '' && $_POST['action'] != 'history_truncate') {
+    if ($sFileName != '' && $aFileInfo['idsfi'] != '' && ($_POST['action'] != 'history_truncate' || $readOnly)) {
         $oVersionStyle = new cVersionFile($aFileInfo['idsfi'], $aFileInfo, $sFileName, $sTypeContent, $cfg, $cfgClient, $db, $client, $area, $frame);
 
         // Init Form variables of SelectBox
@@ -139,8 +144,8 @@ if (!$perm->have_perm_area_action($area, 'style_history_manage')) {
             // Create Textarea and fill it with xml nodes
             if (count($aNodes) > 1) {
                 // If choose xml file read value an set it
-                $sName = $oVersionStyle->getTextBox('stylename', $aNodes['name'], 60);
-                $description = $oVersionStyle->getTextarea('styledesc', (string) $aNodes['desc'], 100, 10);
+                $sName = $oVersionStyle->getTextBox('stylename', $aNodes['name'], 60, $readOnly);
+                $description = $oVersionStyle->getTextarea('styledesc', (string) $aNodes['desc'], 100, 10, '', $readOnly);
                 $sCode = $oVersionStyle->getTextarea('stylecode', (string) $aNodes['code'], 100, 30, 'IdLaycode');
             }
         }
@@ -157,6 +162,9 @@ if (!$perm->have_perm_area_action($area, 'style_history_manage')) {
         $oPage->setEncoding('utf-8');
 
         $oCodeMirrorOutput = new CodeMirror('IdLaycode', 'css', substr(strtolower($belang), 0, 2), true, $cfg, !$bInUse);
+        if($readOnly) {
+            $oCodeMirrorOutput->setProperty("readOnly", "true");
+        }
         $oPage->addScript($oCodeMirrorOutput->renderScript());
 
         if ($sSelectBox != '') {
